requirements.txt
Flask==2.0.2
pymongo==3.12.0
requests==2.26.0
aiohttp==3.8.1
python-dotenv==0.19.1
